refactor and query balances
